Move Parser in GuiConsole
