+++ src/gui/GuiConsole/GuiConsoleMessageParser.h
